+++ trunk/src/gui/canvas.hpp
@@ -56,7 +56,7 @@ public:
 
 private:
     // private constants
-    static const unsigned TILE_POINTEDNESS = 3; // corners of square tiles
+    static const uint16_t TILE_POINTEDNESS = 3; // corners of square tiles
 
     // private data
     static std::vector<Poly> msShapes; // TODO more choices
